img_processing.py
# img_processing.py
import os
from PIL import Image
import img_db
import asyncio

async def process_images(downloaded_images):
    output_dir = "processed_images"
    os.makedirs(output_dir, exist_ok=True)

    db_connection = img_db.connect_to_database()
    valid_image_count = 0

    for i, img_data in enumerate(downloaded_images):
        filename = f"image_{i}.jpg"
        try:
            img = await is_valid_image(img_data)
            if img:
                with img:
                    resized_img = await resize_image(img)
                    resized_img.save(os.path.join(output_dir, filename))

                    img_db.insert_image_metadata(
                        connection=db_connection,
                        url=image_urls[i],
                        filename=filename,
                        width=resized_img.width,
                        height=resized_img.height
                    )
                    valid_image_count += 1
            else:
                print(f"Invalid image data for image {i}")
        except (IOError, OSError) as e:
            print(f"Error processing image {i}: {e}")
        except Exception as e:
            print(f"Unexpected error processing image {i}: {e}")

    db_connection.close()
    print(f"{valid_image_count}/{len(downloaded_images)} images downloaded, resized, and saved successfully!")

async def is_valid_image(img_data):
    # Implementation to validate image data
    pass

async def resize_image(img):
    # Implementation to resize image
    pass