requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
Pillow==9.3.0
psycopg2-binary

